requirements.txt
python-telegram-bot==20.7
psutil==5.9.7
python-nmap==0.7.1
scapy==2.5.0
python-whois==0.9.4
dnspython==2.4.2
requests==2.31.0
aiohttp==3.9.1
speedtest-cli==2.1.3
flask==3.0.0
qrcode[pil]==7.4.2
Pillow==10.2.0
